src/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv
pydantic-settings

langchain
chroma
chromadb
openai

json_repair
huggingface_hub
rarfile

transformers
accelerate
safetensors

torch

--extra-index-url https://download.pytorch.org/whl/cu121